configure, setup: restructure setup, configurtion yml handling.
As part of #13,

This is to ensure the config template is rendered for each asset,
we read in the config .yml files at invocation and render
the config template for each asset.

This ensures the config being applied does not change during runtime,
and that any template variables are rendered based on the asset.