Prevent leak of vault token in readme instructions
Command line arguments are visible to all processes, so by supplying
`--header X-Vault-Token: TOKEN` on the command line will make the vault
token visible to all processes on the system. We don't want this, so use
command substitution.

By using `<( )`, and `--config`, curl will treat the header as a file,
and hence will not be supplied as a command line argument visible to
other processes. `builtin printf` is used so bash ensures its internal
printf is used instead of `/usr/bin/printf`.